v1.72
Re-introduced if-modified-since for 4chan, I removed various unnecessary connections to 4chan in general so downloading may get on with it faster.
I also replaced the string arrays from the downloads on 4chan, the way (mirage?) did it would probably point to a null reference so I just got rid of it and used a list instead. I'm also probably gonna redo the entire program. Probably.
Bug fixes if any.